cli/endpoints/online/custom-container/triton/requirements.txt
numpy
pillow
numpy